pyproject.toml
[project]
name = "goodmorning-api"
version = "0.1.0"
description = "FastAPI application for GoodMorning AI model serving"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "fastapi",
    "uvicorn",
    "torch",
    "torchvision",
    "pillow",
    "requests"
]
